requirements.txt
psutil==5.9.8
PyQt6==6.6.1
scapy==2.5.0
requests==2.31.0
matplotlib==3.8.2
pyqtgraph==0.13.3
python-whois==0.8.0
dnspython==2.5.0 